@@ -1,15 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <Product>Google Translate</Product>
-    <Version>1.0.2</Version>
+    <Version>1.0.3</Version>
     <Description>Multilingual neural machine translation service to translate text, documents and websites from one language into another.</Description>
     <AssemblyName>Apps.GoogleTranslate</AssemblyName>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Blackbird.Applications.Sdk.Common" Version="2.2.0-alpha3" />
+    <PackageReference Include="Blackbird.Applications.Sdk.Common" Version="2.3.2-alpha1" />
+    <PackageReference Include="Blackbird.Applications.SDK.Extensions.FileManagement" Version="1.0.0-alpha2" />
+    <PackageReference Include="Blackbird.Applications.Sdk.Utils" Version="1.0.13" />
     <PackageReference Include="Google.Cloud.Translate.V3" Version="3.2.0" />
     <PackageReference Include="MimeTypes" Version="2.4.1">
       <PrivateAssets>all</PrivateAssets>
