@@ -1,11 +1,11 @@
 ﻿using Blackbird.Applications.Sdk.Common;
-using File = Blackbird.Applications.Sdk.Common.Files.File;
+using Blackbird.Applications.Sdk.Common.Files;
 
 namespace Apps.GoogleTranslate.Models.Requests
 {
     public class TranslateDocumentRequest
     {
-        public File File { get; set; }
+        public FileReference File { get; set; }
 
         [Display("Target language code")]
         public string TargetLanguageCode { get; set; }
