@@ -1,11 +1,11 @@
 ﻿using Blackbird.Applications.Sdk.Common;
-using File = Blackbird.Applications.Sdk.Common.Files.File;
+using Blackbird.Applications.Sdk.Common.Files;
 
 namespace Apps.GoogleTranslate.Models.Responses
 {
     public class TranslateDocumentResponse
     {
-        public File File { get; set; }
+        public FileReference File { get; set; }
 
         [Display("Detected source language")]
         public string DetectedSourceLanguage { get; set; }
